Müşteri ürünleri görebiliyor.
Satıcı ürün ekleyebiliyor.

Müşteri sepet oluşturabiliyor.

Sipariş verilebiliyor.

Stripe üzerinden ödeme yapılabiliyor (test ortamı).

Ödeme başarılıysa sipariş otomatik DELIVERED oluyor.

Admin isterse siparişleri yönetebiliyor.

+++ src/main/java/ramazan/sahin/ecommerce/entity/OrderItem.java
@@ -1,9 +1,10 @@
 package ramazan.sahin.ecommerce.entity;
 
 
+import java.math.BigDecimal;
+
 import jakarta.persistence.*;
 import lombok.*;
-
 @Entity
 @Table(name = "order_items")
 @Data
@@ -28,5 +29,5 @@ public class OrderItem {
     private Integer quantity;
 
     @Column(nullable = false)
-    private Double unitPrice;
+    private BigDecimal unitPrice;
 }
